Create Binary_Search.cpp
To search for an element into some list using the binary search technique, we must ensure that the list is sorted. Binary search looks for a particular item by comparing the middle most item of the collection. If a match occurs, then the index of item is returned. If the middle item is greater than the item, then the item is searched in the sub-array to the left of the middle item. Otherwise, the item is searched for in the sub-array to the right of the middle item. This process continues the sub-array as well until the size of the subarray reduces to zero.